@@ -1,7 +1,10 @@
 source 'http://rubygems.org'
 
-gem 'rails', '4.0.2'
+gem 'rails', '4.1.4'
+gem 'spring', group: :development
 gem 'puma'
+gem 'foreman'
+gem 'sdoc', require: false
 
 #custom gems
 gem 'devise'
@@ -11,24 +14,42 @@ gem 'paypal-sdk-permissions'
 gem 'paymill'
 gem 'money-rails'
 gem 'monetize'
-gem 'paperclip', github: 'thoughtbot/paperclip'
 gem 'aws-sdk'
 gem 'simple_form'
 gem 'agent_orange'
-gem 'eu_central_bank'
 gem 'kaminari'
 gem 'sidekiq'
-gem 'sinatra', '>= 1.3.0', :require => nil
-gem 'whenever', require: false
+gem 'sinatra', '>= 1.3.0', require: nil
 gem 'remotipart', '~> 1.2'
-#test gems
+gem 'rack-affiliates'
+gem 'carrierwave'
+gem 'fog'
+gem 'carrierwave_direct', git: 'https://github.com/Rodeoclash/carrierwave_direct'
+gem 'mini_magick'
+gem 'carrierwave-processing'
+gem "exchange", "~> 1.2.0"
+gem 'redcarpet'
+gem 'maildown'
+gem 'omniauth-createsend'
+gem 'omniauth-mailchimp'
+gem 'tabs', git: 'https://github.com/byterussian/tabs.git', branch: 'customization'
+gem 'email_validator'
+gem 'premailer-rails'
+gem "paranoia", "~> 2.0"
+
+gem 'omniauth-paymill'
+gem 'createsend'
+gem 'gibbon'
+gem 'mandrill-api'
+gem 'quiet_assets', group: :development
+
 group :development, :test do
   gem 'rspec-rails'
   gem 'factory_girl_rails'
   gem 'byebug'
   gem 'guard-rspec', require: false
-  gem 'jasmine'
   gem 'fuubar'
+  gem 'vcr'
 end
 group :test do
   gem 'faker'
@@ -36,6 +57,7 @@ group :test do
   gem 'selenium-webdriver'
   gem 'database_cleaner'
   gem 'webmock'
+  gem 'timecop'
   gem 'sqlite3'
   gem 'simplecov', '~> 0.7.1', require: false
 end
@@ -51,18 +73,23 @@ gem 'turbolinks'
 gem 'jbuilder', '~> 1.2'
 gem 'therubyracer', platforms: :ruby
 gem 'uglifier', '>= 1.3.0'
+gem 'jquery-ui-rails'
+
 # Css gems
 gem "less-rails"
-gem 'sass-rails'
+gem 'sass-rails', github: 'rails/sass-rails'
 gem "bower-rails", "~> 0.7.0"
 gem "font-awesome-rails"
+gem 'entypo-rails'
 
 # Deployment
 group :development do
-  gem 'capistrano'
-  gem 'capistrano-rails', '~> 1.1'
+  gem 'capistrano', '~> 3.2.0'
+  gem 'capistrano-rbenv', '~> 2.0'
   gem 'capistrano-bundler', '~> 1.1.2'
-  gem 'capistrano-sidekiq'
-  gem 'capistrano-rvm'
-  gem 'foreman'
+  gem 'capistrano-rails', '~> 1.1'
+  gem "better_errors"
+  gem 'pry-rails'
+  gem 'guard-livereload', require: false
+  gem 'rb-fsevent'
 end
