@@ -1,44 +1,74 @@
 class Order < ActiveRecord::Base
+  include Unsubscribable
 
- belongs_to :product
- monetize :amount_cents
- monetize :amount_base_cents, :with_currency => :eur
+  STATUS_NAMES = %w(created completed refunded)
+  belongs_to :product
+  belongs_to :user
+  monetize :amount_cents
+  scope :completed, -> { where status: 'completed' }
+  scope :email_starts_with, -> (query) { where("email like ?", "%#{query}%")}
+  validates :status, inclusion: { in: STATUS_NAMES }
+  serialize :payment_details
 
- before_create { self.token = SecureRandom.urlsafe_base64(16)}
- before_create {self.amount_base = self.amount.exchange_to('EUR')}
- after_save :update_metrics
+  before_create do
+    self.token = SecureRandom.urlsafe_base64(16)
+    self.amount_base_cents = self.amount.cents.in(self.amount_currency.downcase.to_sym).to(:eur)
+    self.buyer_accepts_marketing = true
+  end
 
- def unsubscribe_token
-  verifier =  ActiveSupport::MessageVerifier.new(
-    URI::escape(Upandsell::Application.config.secret_key_base))
-  token = verifier.generate("#{self.id}")
-  {order: self.id, signature: token}
-end
+  after_save :send_emails
+  before_save :increment_order_number
 
+  def refund
+    if self.gateway != 'paypal'
+     service = PaymentService.new('paymill')
+   else
+      service = PaymentService.new('paypal')
+   end
 
-def self.is_valid_token?(id, signature)
-  Order.find(id).unsubscribe_token[:signature] == signature
-end
+   service.refund(self.gateway_token)
+   end
 
-def self.unsubscribe(order_id, signature)
-  if self.is_valid_token?(order_id, signature)
-    user = self.find order_id
-    user.update_attribute :email_subscription, false
+   private
+   def increment_order_number
+    if self.status == 'completed' && self.status_was != 'completed'
+     user = User.find(self.product.user_id)
+     prev_order = user.orders
+     .where("status = ? OR status = ?", 'completed', 'refunded')
+     .order("id DESC").first
+     if prev_order
+       prev_number = prev_order.number.to_i
+     else
+      prev_number = 0
+    end
+
+    self.number = prev_number + 1
   end
 end
 
+def self.confirm_unsubscribe(id, type, signature)
+  unsubscribe(id, type, signature)
+end
 private
-def update_metrics
+def send_emails
   if self.status == 'completed' && self.status_was != 'completed'
-   user = User.find(self.product.user_id)
-   UserMailer.delay.bought_email(user, self)
-   if user.email_after_sale
-     UserMailer.delay.sold_email(user, self)
-   end
- end
- if self.status == 'refunded' && self.status_was == 'completed'
-  UserMailer.delay.refund_email(user, self)
-end
-end
+    user = User.find(self.product.user_id)
+    UserMailer.delay.bought_email(user.id, self.id)
+    Metric::Product.new(self.product).record_sale(self.amount_base_cents)
 
+
+    if user.email_after_sale
+      UserMailer.delay.sold_email(user.id, self.id)
+    end
+
+    if self.buyer_accepts_marketing
+      MailingListAddSyncWorker.perform_async(self.id)
+    end
+  end
+
+  if self.status == 'refunded' && self.status_was == 'completed'
+    UserMailer.delay.refund_email(user.id, self.id)
+    Metric::Product.new(self.product).delete_sale(self.amount_base_cents, self.created_at)
+  end
+end
 end
