@@ -1,9 +1,9 @@
-if Rails.env == "production"
-  Sidekiq.configure_server do |config|
-    config.redis = { :url => 'redis://db.upandsell.me:6379', :namespace => 'sidekiq' }
-  end
+redis = Rails.application.secrets.redis
+redis_url = "redis://#{redis['host']}:#{redis['port']}/#{redis['db']}"
 
-  Sidekiq.configure_client do |config|
-    config.redis = { :url => 'redis://db.upandsell.me:6379', :namespace => 'sidekiq' }
-  end
+Sidekiq.configure_server do |config|
+  config.redis = { url: redis_url, namespace: 'sidekiq' }
+end
+Sidekiq.configure_client do |config|
+  config.redis = { url: redis_url, namespace: 'sidekiq' }
 end
