@@ -1,12 +1,14 @@
 // Product landing page related js
 //= require jquery
 //= require jquery_ujs
+//= require jquery-ui
+//= require excanvas.compiled.js
+//= require canvg.js
+//= require polyclip.js
 //= require bootstrap/dist/js/bootstrap
+//= require jquery.scrollTo/jquery.scrollTo.js
 //= require jquery.validation/jquery.validate.js
-//= require slimScroll/jquery.slimscroll
-//= require custombox/src/jquery.custombox.js
 //= require jquery.payment/lib/jquery.payment.js
-//= require jquery.transit/jquery.transit
 //= require product_page
 $(document).ready(function() {
 
