@@ -1,14 +1,19 @@
 (function ($, Up, undefined) {
 
   Up.init = function() {
-
     if ($("#js-message").text().length > 20) { Up.alert.open();}
-    $("select").selectpicker({style: 'btn btn-primary', menuStyle: 'dropdown-inverse'});
+    $("select").selectpicker({style: 'btn btn-primary', menuStyle: 'dropdown-default'});
     $(".select").removeClass('form-control');
-    $("[data-toggle='switch']").wrap('<div class="switch" />').parent().bootstrapSwitch();
-    var page = Utils.getPage();
 
-    try {
+    if($("select").data('selected')){
+     $('.selectpicker').selectpicker('val', $("select").data('selected'));
+   }
+
+
+   $("[data-toggle='switch']").wrap('<div class="switch" />').parent().bootstrapSwitch();
+   var page = Utils.getPage();
+console.log(page);
+   try {
      window[page.controller.capitalize()][page.action.capitalize()]();
    } catch(e) {
   // fail silently
@@ -20,7 +25,7 @@ Up.alert = {
   open: function(message) {
     $('#js-message').text(message);
     $('.animated-alert').animate({
-     top: 60
+     top: 55
    }, 500);
     setTimeout(Up.alert.close, 4500);
   },
