@@ -1,9 +1,9 @@
 class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
-  before_filter :configure_permitted_parameters, if: :devise_controller?
   layout 'site'
-  protect_from_forgery with: :exception
+  protect_from_forgery
+  before_action :configure_permitted_parameters, if: :devise_controller?
 
   def after_sign_in_path_for(user)
    user_root_path
@@ -11,7 +11,8 @@ class ApplicationController < ActionController::Base
 
  protected
  def configure_permitted_parameters
-  devise_parameter_sanitizer.for(:sign_up) << :name
+  devise_parameter_sanitizer.for(:sign_up) { |u| u.permit(:name, :email,
+    :password, :password_confirmation, :referer_id) }
 end
 
 end
