@@ -2,7 +2,7 @@ class ConfirmationsController < Devise::ConfirmationsController
 
   private
   def after_confirmation_path_for(resource_name, resource)
-    user_settings_setup_path
+    user_setup_path
   end
 
 end
