@@ -1,7 +1,6 @@
 class EmailsController < ApplicationController
 
   def unsubscribe_user
-
   end
 
   def confirm_unsubscribe_user
@@ -15,23 +14,19 @@ class EmailsController < ApplicationController
   end
 
   def unsubscribe_order
-    response = Order.unsubscribe(params[:order], params[:signature])
-    if response
-     return
-   end
-   render json: {status: :not_found}
- end
-
- def confirm_unsubscribe_order
-  response = Order.unsubscribe(params[:order], params[:signature])
-  if response
-    render "emails/unsubscribed"
-    return
   end
 
-  render 'unsubscribe_order' and return
-end
+  def confirm_unsubscribe_order
+    response = Order.confirm_unsubscribe(params[:order], params[:type], params[:signature])
+    if response
+      MailingListRemoveSyncWorker.perform_async(params[:order])
+      render "emails/unsubscribed"
+      return
+    end
 
-def unsubscribed
-end
+    render 'unsubscribe_order' and return
+  end
+
+  def unsubscribed
+  end
 end
