@@ -1,4 +1,4 @@
 require 'aws-sdk'
-Rails.configuration.aws = YAML.load_file(Rails.root.join("config/aws.yml")).symbolize_keys[Rails.env.to_sym]
+Rails.configuration.aws = Rails.application.secrets.aws
 AWS.config(logger: Rails.logger)
 AWS.config(Rails.configuration.aws)
