@@ -2,12 +2,19 @@ include ActionDispatch::TestProcess
 
 FactoryGirl.define do
   factory :order do
-    email "dddd@cldlf.com"
+    email "ddd@cldlf.com"
     amount_cents 300
     amount_base_cents 150
     amount_currency 'USD'
-    status 'completed'
+    status 'created'
+    token 'tok_a428697472ff0fd78f7a'
+    gateway 'paypal'
+    user
     product
+  end
+
+  factory :order_completed, parent: :order do
+    status 'completed'
   end
 
   factory :order_yankee, parent: :order do
